requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMuPDF==1.23.8
python-docx==1.1.0
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
pytest==7.4.3
Werkzeug==3.0.1
